Create backup_mul_netmiko
This script is to handle multiple cisco version devices for collecting the Configuration. 
Script creates a new folder according to the date stamp. Reads the input file for various switches which has been created as a Dictionary as per Netmiko suggestions.
On each for loop script verifies the type of device and runs through particular parts of code. I have created a CRON job for this so which required an extra command to enter the folder when CRON job calls the script from the root folder. 
Finally when SSH to the device it scraps the config on the devices and reverts back to unique file name created with Name of the device and timestamp.